==> rnaseq/07_WGCNA_GO.ipynb <==
## Set Library Path

.libPaths("/home/vhaghani/anaconda3/envs/jupyter_nb_R4.4.2/lib/R/library")

## Load Libraries

library(dplyr)
library(tidyr)
library(tidyverse)
library(enrichR)
library(openxlsx)
library(ggplot2)

## Load Data

# Read in modules
module_membership <- read.delim("06_WGCNA/filtered_module_membership.txt")

# View data
head(module_membership)

## Separate Genes Per Module

# View modules
modules <- unique(module_membership$Module)

print(modules)

# Iterate over each module
for (module in modules) {
  # Filter data for the current module
  module_data <- module_membership[module_membership$Module == module, ]
  
  # Extract genes from external_gene_name column
  genes <- module_data$external_gene_name
  
  # Remove NA values
  genes <- genes[!is.na(genes)]
  
  # Create a data frame with a single column containing genes
  genes_df <- data.frame(Genes = genes)
  
  # Define the filename for the CSV
  filename <- paste0("06_WGCNA/", module, "_genes.csv")

  # Rename the column
  colnames(genes_df) <- module
    
  # Write the genes to a CSV file
  write.csv(genes_df, file = filename, row.names = FALSE)
}

# List to store genes for each module
gene_lists <- list()

# Iterate over each module
for (module in modules) {
  # Define the filename for the CSV
  filename <- paste0("06_WGCNA/", module, "_genes.csv")
  
  # Read CSV file
  module_genes <- read.csv(filename, header = FALSE)$V1
  
  # Store genes in the list
  gene_lists[[module]] <- module_genes
}

# Find the maximum length among all gene lists
max_length <- max(sapply(gene_lists, length))

# Pad shorter gene lists with NA values to make all lists equal length
for (module in modules) {
  diff_length <- max_length - length(gene_lists[[module]])
  if (diff_length > 0) {
    gene_lists[[module]] <- c(gene_lists[[module]], rep(NA, diff_length))
  }
}

# Combine gene lists into a data frame
merged_df <- as.data.frame(gene_lists)

# Rename columns with module names
colnames(merged_df) <- modules

# Remove the first row that has module names
merged_df <- merged_df[-1, ]

# Print the first few rows of the merged data frame
head(merged_df)

# Save all modules and genes into a dataframe
write.csv(merged_df, file = "06_WGCNA/all_genes_per_module.csv", row.names = FALSE)

## Carry Out Gene Ontology with EnrichR

# Iterate over each module
for (module in modules) {
  tryCatch({
    # Perform the enrichR analysis on the gene list for the current module
    enrichr_results <- enrichr(merged_df[[module]], c("GO_Biological_Process_2023",
                                                     "GO_Cellular_Component_2023",
                                                     "GO_Molecular_Function_2023",
                                                     "KEGG_2019_Mouse",
                                                     "Panther_2016",
                                                     "Reactome_2016",
                                                     "RNA-Seq_Disease_Gene_and_Drug_Signatures_from_GEO"))
    
    # Check if enrichr_results is empty
    if (length(enrichr_results) == 0) {
      cat("No results for module", module, "\n")
      next
    }
    
    # Save Enrichr outputs
    wb <- createWorkbook()
    
    for (i in seq_along(enrichr_results)) {
      # Extract the data frame from the list
      df <- enrichr_results[[i]]
      
      # Check if the data frame is empty
      if (nrow(df) == 0) {
        cat("Empty data frame for", names(enrichr_results)[i], "in module", module, "\n")
        next
      }
      
      # Define the original sheet name
      original_sheet_name <- names(enrichr_results)[i]
      
      # Modify the sheet name if it's specifically "RNA-Seq_Disease_Gene_and_Drug_Signatures_from_GEO"
      sheet_name <- if (original_sheet_name == "RNA-Seq_Disease_Gene_and_Drug_Signatures_from_GEO") {
        "RNAseq_DiseaseGene_DrugSigs_GEO"
      } else {
        original_sheet_name
      }
      
      # Add the data frame as a new sheet in the Excel workbook
      addWorksheet(wb, sheet_name)
      writeData(wb, sheet = sheet_name, x = df)
    }
    
    # Save the Excel workbook
    saveWorkbook(wb, paste0("06_WGCNA/", module, "_enrichr_results.xlsx"), overwrite = TRUE)
    
    # Function to plot and save the results
    plot_and_save <- function(df, filename, title) {
      if (nrow(df) == 0) {
        cat("Empty data frame for", title, "in module", module, "\n")
        return()
      }
      pdf(filename, height = 7, width = 15)
      print(plotEnrich(df, showTerms = 25, numChar = 75, y = "Count", orderBy = "P.value") + ggtitle(title))
      dev.off()
    }
    
    # Plot and save Enrichr results
    plot_and_save(enrichr_results$GO_Biological_Process_2023, 
                  paste0("06_WGCNA/", module, "_GO_Biological_Process_2023.pdf"), 
                  paste("GO_Biological_Process_2023 for", module, "module"))
    
    plot_and_save(enrichr_results$GO_Cellular_Component_2023, 
                  paste0("06_WGCNA/", module, "_GO_Cellular_Component_2023.pdf"), 
                  paste("GO_Cellular_Component_2023 for", module, "module"))
    
    plot_and_save(enrichr_results$GO_Molecular_Function_2023, 
                  paste0("06_WGCNA/", module, "_GO_Molecular_Function_2023.pdf"), 
                  paste("GO_Molecular_Function_2023 for", module, "module"))
    
    plot_and_save(enrichr_results$KEGG_2019_Mouse, 
                  paste0("06_WGCNA/", module, "_KEGG_2019_Mouse.pdf"), 
                  paste("KEGG_2019_Mouse for", module, "module"))
    
    plot_and_save(enrichr_results$Panther_2016, 
                  paste0("06_WGCNA/", module, "_Panther_2016.pdf"), 
                  paste("Panther_2016 for", module, "module"))
    
    plot_and_save(enrichr_results$Reactome_2016, 
                  paste0("06_WGCNA/", module, "_Reactome_2016.pdf"), 
                  paste("Reactome_2016 for", module, "module"))
    
    plot_and_save(enrichr_results$`RNA-Seq_Disease_Gene_and_Drug_Signatures_from_GEO`, 
                  paste0("06_WGCNA/", module, "_RNA-Seq_Disease_Gene_and_Drug_Signatures_from_GEO.pdf"), 
                  paste("RNA-Seq_Disease_Gene_and_Drug_Signatures_from_GEO for", module, "module"))
    
  }, error = function(e) {
    cat("Error occurred for module", module, ": ", conditionMessage(e), "\n")
    # Log the error to a file for further inspection
    write(paste("Error occurred for module", module, ": ", conditionMessage(e), "\n"), file = "error_log.txt", append = TRUE)
    # Continue to the next module
    next
  })
}
